python/requirements.txt
Flask==1.1.2
oauthlib==3.1.0
requests_oauthlib==1.3.0
python-decouple==3.3
Flask-MongoEngine==1.0.0